requirements.txt
dj-database-url==0.5.0
Django==3.1
gunicorn==20.0.4
psycopg2==2.8.5
pytest-django==3.9.0
pytest-mock==3.2.0
flake8
whitenoise==5.2.0
djangorestframework==3.11.1
usaddress==0.5.10
parserator
six
